refactor: extract ProfileHandler from MCPServerV2
Move 10 profile management methods (CRUD, parsing, directory
lookup) into dedicated ProfileHandler. Server retains thin
delegation wrappers and MCP tool registration.
