+++ src/llm_orc/mcp/server.py
@@ -19,6 +19,7 @@ from llm_orc.core.config.config_manager import ConfigurationManager
 from llm_orc.core.config.ensemble_config import EnsembleLoader, assert_no_cycles
 from llm_orc.core.execution.artifact_manager import ArtifactManager
 from llm_orc.mcp.handlers.help_handler import HelpHandler
+from llm_orc.mcp.handlers.profile_handler import ProfileHandler
 
 if TYPE_CHECKING:
     from llm_orc.core.execution.ensemble_execution import EnsembleExecutor
@@ -132,6 +133,7 @@ class MCPServerV2:
         self.artifact_manager = ArtifactManager()
         self._executor = executor  # Lazily created if None
         self._help_handler = HelpHandler()
+        self._profile_handler = ProfileHandler(self.config_manager)
         self._mcp = FastMCP("llm-orc")
         self._setup_resources()
         self._setup_tools()
@@ -2074,192 +2076,20 @@ class MCPServerV2:
     # =========================================================================
 
     async def _list_profiles_tool(self, arguments: dict[str, Any]) -> dict[str, Any]:
-        """List model profiles.
-
-        Args:
-            arguments: Tool arguments including optional provider filter.
-
-        Returns:
-            List of profiles.
-        """
-        import yaml
-
-        provider_filter = arguments.get("provider")
-        profiles: list[dict[str, Any]] = []
-
-        # Get profiles from configured directories
-        profiles_dirs = self.config_manager.get_profiles_dirs()
-
-        for dir_path in profiles_dirs:
-            profiles_dir = Path(dir_path)
-            if not profiles_dir.exists():
-                continue
-
-            for yaml_file in profiles_dir.glob("*.yaml"):
-                try:
-                    content = yaml_file.read_text()
-                    data = yaml.safe_load(content) or {}
-                    profile_provider = data.get("provider", "")
-
-                    # Apply filter if specified
-                    if provider_filter and profile_provider != provider_filter:
-                        continue
-
-                    profiles.append(
-                        {
-                            "name": data.get("name", yaml_file.stem),
-                            "provider": profile_provider,
-                            "model": data.get("model", ""),
-                            "path": str(yaml_file),
-                        }
-                    )
-                except Exception:
-                    continue
-
-        return {"profiles": profiles}
-
-    def _get_local_profiles_dir(self) -> Path:
-        """Get the local profiles directory for writing.
-
-        Returns:
-            Path to local profiles directory.
-
-        Raises:
-            ValueError: If no profiles directory is configured.
-        """
-        profiles_dirs = self.config_manager.get_profiles_dirs()
-        for dir_path in profiles_dirs:
-            path = Path(dir_path)
-            if ".llm-orc" in str(path) and "library" not in str(path):
-                return path
-        if profiles_dirs:
-            return Path(profiles_dirs[0])
-        raise ValueError("No profiles directory configured")
+        """List model profiles."""
+        return await self._profile_handler.list_profiles(arguments)
 
     async def _create_profile_tool(self, arguments: dict[str, Any]) -> dict[str, Any]:
-        """Create a new profile.
-
-        Args:
-            arguments: Tool arguments.
-
-        Returns:
-            Creation result.
-        """
-        import yaml
-
-        name = arguments.get("name")
-        provider = arguments.get("provider")
-        model = arguments.get("model")
-
-        if not name:
-            raise ValueError("name is required")
-        if not provider:
-            raise ValueError("provider is required")
-        if not model:
-            raise ValueError("model is required")
-
-        local_dir = self._get_local_profiles_dir()
-        target_file = local_dir / f"{name}.yaml"
-        if target_file.exists():
-            raise ValueError(f"Profile '{name}' already exists")
-
-        # Build profile data
-        profile_data: dict[str, Any] = {
-            "name": name,
-            "provider": provider,
-            "model": model,
-        }
-        if arguments.get("system_prompt"):
-            profile_data["system_prompt"] = arguments["system_prompt"]
-        if arguments.get("timeout_seconds"):
-            profile_data["timeout_seconds"] = arguments["timeout_seconds"]
-        if arguments.get("temperature") is not None:
-            profile_data["temperature"] = arguments["temperature"]
-        if arguments.get("max_tokens") is not None:
-            profile_data["max_tokens"] = arguments["max_tokens"]
-
-        # Write file
-        local_dir.mkdir(parents=True, exist_ok=True)
-        yaml_content = yaml.safe_dump(profile_data, default_flow_style=False)
-        target_file.write_text(yaml_content)
-
-        return {"created": True, "path": str(target_file)}
+        """Create a new profile."""
+        return await self._profile_handler.create_profile(arguments)
 
     async def _update_profile_tool(self, arguments: dict[str, Any]) -> dict[str, Any]:
-        """Update an existing profile.
-
-        Args:
-            arguments: Tool arguments.
-
-        Returns:
-            Update result.
-        """
-        import yaml
-
-        name = arguments.get("name")
-        changes = arguments.get("changes", {})
-
-        if not name:
-            raise ValueError("name is required")
-
-        # Find profile file
-        profiles_dirs = self.config_manager.get_profiles_dirs()
-        profile_file = None
-
-        for dir_path in profiles_dirs:
-            path = Path(dir_path) / f"{name}.yaml"
-            if path.exists():
-                profile_file = path
-                break
-
-        if not profile_file:
-            raise ValueError(f"Profile '{name}' not found")
-
-        # Load and update
-        content = profile_file.read_text()
-        data = yaml.safe_load(content) or {}
-        data.update(changes)
-
-        # Write back
-        yaml_content = yaml.safe_dump(data, default_flow_style=False)
-        profile_file.write_text(yaml_content)
-
-        return {"updated": True, "path": str(profile_file)}
+        """Update an existing profile."""
+        return await self._profile_handler.update_profile(arguments)
 
     async def _delete_profile_tool(self, arguments: dict[str, Any]) -> dict[str, Any]:
-        """Delete a profile.
-
-        Args:
-            arguments: Tool arguments.
-
-        Returns:
-            Deletion result.
-        """
-        name = arguments.get("name")
-        confirm = arguments.get("confirm", False)
-
-        if not name:
-            raise ValueError("name is required")
-        if not confirm:
-            raise ValueError("Confirmation required to delete profile")
-
-        # Find profile file
-        profiles_dirs = self.config_manager.get_profiles_dirs()
-        profile_file = None
-
-        for dir_path in profiles_dirs:
-            path = Path(dir_path) / f"{name}.yaml"
-            if path.exists():
-                profile_file = path
-                break
-
-        if not profile_file:
-            raise ValueError(f"Profile '{name}' not found")
-
-        # Delete
-        profile_file.unlink()
-
-        return {"deleted": True, "name": name}
+        """Delete a profile."""
+        return await self._profile_handler.delete_profile(arguments)
 
     # =========================================================================
     # Artifact Management Tool Implementations
@@ -3015,75 +2845,8 @@ if __name__ == "__main__":
         return sorted(available_models)[:5]  # Return top 5
 
     def _get_all_profiles(self) -> dict[str, dict[str, Any]]:
-        """Get all profiles as a dict keyed by name.
-
-        Returns:
-            Dict mapping profile name to profile config.
-        """
-        profiles: dict[str, dict[str, Any]] = {}
-
-        for dir_path in self.config_manager.get_profiles_dirs():
-            profile_dir = Path(dir_path)
-            if not profile_dir.exists():
-                continue
-
-            for yaml_file in profile_dir.glob("*.yaml"):
-                self._load_profiles_from_file(yaml_file, profiles)
-
-        return profiles
-
-    def _load_profiles_from_file(
-        self, yaml_file: Path, profiles: dict[str, dict[str, Any]]
-    ) -> None:
-        """Load profiles from a YAML file into the profiles dict.
-
-        Args:
-            yaml_file: Path to YAML file.
-            profiles: Dict to populate with profiles.
-        """
-        try:
-            import yaml
-
-            with open(yaml_file) as f:
-                data = yaml.safe_load(f) or {}
-
-            self._parse_profile_data(data, profiles)
-        except Exception:
-            pass
-
-    def _parse_profile_data(
-        self, data: dict[str, Any], profiles: dict[str, dict[str, Any]]
-    ) -> None:
-        """Parse profile data from various YAML formats.
-
-        Args:
-            data: Parsed YAML data.
-            profiles: Dict to populate with profiles.
-        """
-        if "model_profiles" in data:
-            self._parse_dict_format_profiles(data["model_profiles"], profiles)
-        elif "profiles" in data:
-            self._parse_list_format_profiles(data["profiles"], profiles)
-        elif "name" in data:
-            profiles[data["name"]] = data
-
-    def _parse_dict_format_profiles(
-        self, model_profiles: dict[str, Any], profiles: dict[str, dict[str, Any]]
-    ) -> None:
-        """Parse dict format: model_profiles: {name: {config...}}."""
-        for name, config in model_profiles.items():
-            if isinstance(config, dict):
-                config["name"] = name
-                profiles[name] = config
-
-    def _parse_list_format_profiles(
-        self, profile_list: list[dict[str, Any]], profiles: dict[str, dict[str, Any]]
-    ) -> None:
-        """Parse list format: profiles: [{name: ..., ...}]."""
-        for p in profile_list:
-            name = p.get("name", "")
-            if name:
-                profiles[name] = p
+        """Get all profiles as a dict keyed by name."""
+        return self._profile_handler.get_all_profiles()
 
     async def _help_tool(self, arguments: dict[str, Any]) -> dict[str, Any]:
         """Get help documentation.
